Companion Commit to @fossdd's Index Enhancements - Remove no-longer-needed ".citation-needed" styles in styes.css:
The [How?] text for which this style was needed,
has now been replaced with the proper instructions.

- Fix button proportions in "Source Code" section:

The "SourcHut" button was roughly equal the length of the "GitHub" button,
but not quite, which looked wrong.
This fix makes it Actually ⅔ of the column wide
(with the "GitHub" button taking the remaining ⅓),
as originally intended.

This size difference is
because the Tenacity Team wants to emphasize SourceHut more than GitHub.

- Adjust proportions in the bottom row of the "Community" section:

As the text in the 2nd button has been shortened
from "Join #tenacity IRC channel (web, /join #tenacity)" to "Web Chat",
both buttons have now been made equally wide.

- Shorten the Button Grids:

With the aforementioned shortened text,
the button grids now have a slightly shorter maximum width
to keep them looking right.

- Rearrange "button.css" to be more clear:

Previously, the ".header-button" styles have been hidden
in the middle of the file. Now they're more clearly separated to the end.

- Elongate the "Download" button slightly:

The left and right padding of the Download Button
has been increased from 1em to 2em
because the button appeared too short otherwise.

- Add padding before and after the main text, as suggested by @fossdd:

> Things I noticed that you can fix:
>   [...]
> - More space above and after the text

Signed-off-by: Sosasees <sosasees@protonmail.com>