Improved Layout
- Added Top Buttons Bar

with the 3 most important links on the page

- Moved the "Wall of Buttons" to the bottom of the page

The most important links are now also in the new "Top Buttons Bar"

- Made the "Features List" into a responsive multi-column layout

I don't really know why, but this makes the Tenacity program
feel a lot more impressive.
Thank @bitsper2nd for suggesting this change:
https://github.com/tenacityteam/tenacity/discussions/487#discussioncomment-1195351

- Restricted the width of the screenshot at the beginning

@bitsper2nd complained about Too Much Whitespace
in the just-linked comment thread.

There's never actually "Too much Whitespace" on websites,
but only the perception thereof when some elements
have much more whitespace than others.

I found that on this webpage, this perception came mainly
from the screenshot being way too wide compared to the rest of the page.
So I shrunk it to be only as wide as the rest of the page.

- Removed the Gradient Background at the beginning of the page

It is now default-colored like the rest of the page

- Added @semisol's Tenacity Mascot, Key the Cat

- Darkened the Address Bar when in Chrome for Android

compared to my previous commit, where it was not darkened enough

Signed-off-by: Sosasees <sosasees@protonmail.com>