Prototype a new "Community & Source Code" page
This section that was previously on the homepage
has now been moved to a subpage
so that it can have more breathing room
and the elements can be much better explained.

This is the first (Minimum Viable, except for that it already has some CSS)
version of the page, and has the footer missing.
I have not yet developed it further so that I can get feedback first.

Signed-off-by: Sosasees <sosasees@protonmail.com>